feat(layout): promote landing footer to the global Footer + FR/EN on mobile
* src/components/layout/Footer.tsx is rewritten with the rich 4-column
  layout that used to live as LandingFooter inside HomePage: brand +
  tagline / Product / Science / Project columns, full-bleed width, and
  the "© year · Zoned" + "Built in public · GitHub" + version row at
  the bottom. Library, Plans, every detail page and the rest of the
  app now share the same footer, which is what was asked for.
* HomePage drops its inline LandingFooter / FooterColumn /
  APP_VERSION definitions; the page now ends with the final CTA and
  lets App.tsx render the global footer underneath.
* App.tsx no longer carries the CUSTOM_FOOTER_ROUTES = ["/"] exception
  since the home and the rest of the app render the same footer.
  FULLSCREEN_ROUTES (/routes) still suppresses the footer for the map
  editor.
* TopBar mobile: the FR/EN switch sits in the right action group
  again (between Search and the theme toggle) — it used to only exist
  on desktop, leaving mobile users with no way to change language
  without opening the menu sheet.
